Expose RunScanner flowcell geometry functions
Rather than a single `flowcell_geometry` function, this exposes the underlying
functions to generate flowcell geometry information without an active
connection to RunScanner based on the information being fetched from
RunScanner:

- `runscanner::flowcell_geometry` - Generate a flowcell geometry for the number of lanes specified depending on whether they are joined together or separate.
- `runscanner::is_workflow_type_joined` Check if a sequencing workflow type has a flow cell where all the lanes are joined or not.
